requirements.txt
matplotlib==3.2.2
numpy==1.22.3
Pillow==9.2.0
torch==1.11.0
torchvision==0.11.2+cu113
